scanner: execute syft scan in chroot environment
When scanning a directory, the syft scanner resolves symlinks inside the
target directory in the context of the host filesystem. However, for our
purposes, the target filesystem is entirely self-contained. In lieu of
syft changes, we can temporarily hack around this limitation by
executing the scanner in a chroot-ed environment.

Signed-off-by: Justin Chadwell <me@jedevc.com>